% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{bleach}
\alias{bleach}
\title{bleach colors by blending toward white}
\usage{
bleach(hex, lighten = 0.7, mute = 0.9)
}
\arguments{
\item{hex}{Character vector of hex colors.}

\item{amount}{Fraction \link{0,1}, 0 = no change, 1 = full white.}
}
\value{
bleached hex colors.
}
\description{
bleach colors by blending toward white
}
