% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{continuous_pal}
\alias{continuous_pal}
\title{Generate continuous color gradient}
\usage{
continuous_pal(cols, n = 10)
}
\arguments{
\item{cols}{Vector of base colors.}

\item{n}{Number of gradient colors.}
}
\value{
Character vector of interpolated colors.
}
\description{
Generate continuous color gradient
}
