% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{coral_palettes}
\alias{coral_palettes}
\title{Coral palettes}
\format{
An object of class \code{list} of length 12.
}
\usage{
coral_palettes
}
\description{
Coral palettes
}
\keyword{datasets}
