% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{darken}
\alias{darken}
\title{Darken colors by blending toward black}
\usage{
darken(hex, amount = 0.2)
}
\arguments{
\item{hex}{Character vector of hex colors.}

\item{amount}{Fraction \link{0,1}, 0 = no change, 1 = full black.}
}
\value{
Darkened hex colors.
}
\description{
Darken colors by blending toward black
}
