% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{diverging_pal}
\alias{diverging_pal}
\title{Select diverging colors spaced apart in Lab space}
\usage{
diverging_pal(pal, n = 7)
}
\arguments{
\item{pal}{Vector of base colors.}

\item{n}{Number of colors to return.}
}
\value{
Character vector of diverging colors.
}
\description{
Select diverging colors spaced apart in Lab space
}
