% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{lighten}
\alias{lighten}
\title{Lighten colors by blending toward white}
\usage{
lighten(hex, amount = 0.2)
}
\arguments{
\item{hex}{Character vector of hex colors.}

\item{amount}{Fraction \link{0,1}, 0 = no change, 1 = full white.}
}
\value{
Lightened hex colors.
}
\description{
Lighten colors by blending toward white
}
