% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{plot_palette}
\alias{plot_palette}
\title{Quick base R palette bar plot}
\usage{
plot_palette(hex)
}
\arguments{
\item{hex}{Character vector of hex colors.}
}
\value{
Bar plot of colors.
}
\description{
Quick base R palette bar plot
}
