% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{plot_palette_plotly}
\alias{plot_palette_plotly}
\title{Interactive palette plot with Plotly}
\usage{
plot_palette_plotly(hex)
}
\arguments{
\item{hex}{Character vector of hex colors.}
}
\value{
Plotly bar chart of colors with hover hex.
}
\description{
Interactive palette plot with Plotly
}
