% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colors.R
\name{show_colors}
\alias{show_colors}
\title{Display colors in a grid using ggplot2}
\usage{
show_colors(
  colours,
  labels = TRUE,
  label_size = 5,
  label_color = "white",
  ncol = NULL,
  nrow = NULL,
  title = NULL,
  title_size = 14,
  title_color = "black"
)
}
\arguments{
\item{colours}{Character vector of hex colors to display.}

\item{labels}{Logical, whether to show color codes as text labels. Default \code{TRUE}.}

\item{label_size}{Numeric, size of the label text.}

\item{label_color}{Color for the label text (default \code{"white"}).}

\item{ncol}{Number of columns in the grid. If \code{NULL} will be calculated automatically.}

\item{nrow}{Number of rows in the grid. If \code{NULL} will be calculated automatically.}

\item{title}{Optional plot title text.}

\item{title_size}{Numeric, size of the title text.}

\item{title_color}{Color of the title text.}
}
\value{
A \code{ggplot} object showing the color swatches.
}
\description{
\code{show_colors()} creates a simple color swatch grid with optional hex labels and optional title.
}
\examples{
\dontrun{
show_colors(c("#0570c2", "#F9B908", "#ad8bb8", "#009382"), title = "My Palette")
}

}
